feat: multi-cube search router
